+++ src/mkvtoolnix-gui/main_window/preferences_dialog.cpp
@@ -124,6 +124,7 @@ PreferencesDialog::PreferencesDialog(QWidget *parent,
   ui->cbMAttachmentAlwaysSkipForExistingName->setChecked(m_cfg.m_mergeAttachmentsAlwaysSkipForExistingName);
 
   setupDeriveForcedDisplayFlagSubtitles();
+  setupDeriveHearingImpairedFlag();
   setupFileColorsControls();
   setupProcessPriority();
   setupPlaylistScanningPolicy();
@@ -518,6 +519,14 @@ PreferencesDialog::setupToolTips() {
                    .arg(QYH("Depending on the setting below this language can also be used if the language in the source file is 'undetermined' ('und').")));
 
   auto text = Q("<p>%1 %2</p>")
+    .arg(QYH("Audio and subtitle files often contain the words 'cc' or 'sdh' in their file name to signal that they're intended for hearing impaired people."))
+    .arg(QYH("The GUI can set the 'hearing impaired' flag for such tracks if the file name matches this regular expression."));
+
+  Util::setToolTip(ui->cbMDeriveHearingImpairedFlag,   text);
+  Util::setToolTip(ui->leMDeriveHearingImpairedFlagRE, text);
+  Util::setToolTip(ui->pbMDeriveHearingImpairedFlagRERevert, QY("Revert the entry to its default value."));
+
+  text = Q("<p>%1 %2</p>")
     .arg(QYH("Subtitle files often contain the word 'forced' in their file name to signal that they're intended for 'forced display' only (e.g. when they speak Elfish in 'Lord of the Rings')."))
     .arg(QYH("The GUI can set the 'forced display' flag for such tracks if the file name matches this regular expression."));
 
@@ -645,6 +654,8 @@ PreferencesDialog::setupConnections() {
   connect(ui->cbMEnableMuxingTracksByLanguage,            &QCheckBox::toggled,                                           ui->tbMEnableMuxingTracksByLanguage,  &QLabel::setEnabled);
 
   connect(ui->pbMDeriveTrackLanguageRevertBoundaryChars,  &QPushButton::clicked,                                         this,                                 &PreferencesDialog::revertDeriveTrackLanguageFromFileNameChars);
+  connect(ui->pbMDeriveHearingImpairedFlagRERevert,       &QPushButton::clicked,                                         this,                                 &PreferencesDialog::revertDeriveHearingImpairedFlagRE);
+  connect(ui->cbMDeriveHearingImpairedFlag,               &QCheckBox::toggled,                                           this,                                 &PreferencesDialog::enableDeriveHearingImpairedFlagControls);
   connect(ui->pbMDeriveForcedDisplayFlagSubtitlesRERevert, &QPushButton::clicked,                                        this,                                 &PreferencesDialog::revertDeriveForcedDisplayFlagSubtitlesRE);
   connect(ui->cbMDeriveForcedDisplayFlagSubtitles,         &QCheckBox::toggled,                                          this,                                 &PreferencesDialog::enableDeriveForcedDisplayFlagSubtitlesControls);
 
@@ -959,7 +970,11 @@ PreferencesDialog::setupTabPositions() {
 }
 
 void
+PreferencesDialog::setupDeriveHearingImpairedFlag() {
+  ui->cbMDeriveHearingImpairedFlag->setChecked(m_cfg.m_deriveHearingImpairedFlagFromFileNames);
+  ui->leMDeriveHearingImpairedFlagRE->setText(m_cfg.m_regexForDerivingHearingImpairedFlagFromFileNames);
 
+  enableDeriveHearingImpairedFlagControls();
 }
 
 void
@@ -971,6 +986,13 @@ PreferencesDialog::setupDeriveForcedDisplayFlagSubtitles() {
 }
 
 void
+PreferencesDialog::enableDeriveHearingImpairedFlagControls() {
+  auto enable = ui->cbMDeriveHearingImpairedFlag->isChecked();
+
+  ui->leMDeriveHearingImpairedFlagRE->setEnabled(enable);
+  ui->pbMDeriveHearingImpairedFlagRERevert->setEnabled(enable);
+}
+
 void
 PreferencesDialog::enableDeriveForcedDisplayFlagSubtitlesControls() {
   auto enable = ui->cbMDeriveForcedDisplayFlagSubtitles->isChecked();
@@ -1216,6 +1238,8 @@ PreferencesDialog::save() {
   m_cfg.m_defaultVideoTrackLanguage                           = ui->ldwMDefaultVideoTrackLanguage->language();
   m_cfg.m_defaultSubtitleTrackLanguage                        = ui->ldwMDefaultSubtitleTrackLanguage->language();
   m_cfg.m_whenToSetDefaultLanguage                            = static_cast<Util::Settings::SetDefaultLanguagePolicy>(ui->cbMWhenToSetDefaultLanguage->currentData().toInt());
+  m_cfg.m_deriveHearingImpairedFlagFromFileNames              = ui->cbMDeriveHearingImpairedFlag->isChecked();
+  m_cfg.m_regexForDerivingHearingImpairedFlagFromFileNames    = ui->leMDeriveHearingImpairedFlagRE->text();
   m_cfg.m_deriveSubtitlesForcedFlagFromFileNames              = ui->cbMDeriveForcedDisplayFlagSubtitles->isChecked();
   m_cfg.m_regexForDerivingSubtitlesForcedFlagFromFileNames    = ui->leMDeriveForcedDisplayFlagSubtitlesRE->text();
   m_cfg.m_defaultSubtitleCharset                              = ui->cbMDefaultSubtitleCharset->currentData().toString();
@@ -1445,6 +1469,10 @@ PreferencesDialog::revertDeriveTrackLanguageFromFileNameChars() {
 }
 
 void
+PreferencesDialog::revertDeriveHearingImpairedFlagRE() {
+  ui->leMDeriveHearingImpairedFlagRE->setText(Util::Settings::defaultRegexForDerivingHearingImpairedFlagFromFileName());
+}
+
 void
 PreferencesDialog::revertDeriveForcedDisplayFlagSubtitlesRE() {
   ui->leMDeriveForcedDisplayFlagSubtitlesRE->setText(Util::Settings::defaultRegexForDerivingForcedDisplayFlagForSubtitlesFromFileName());
@@ -1479,6 +1507,22 @@ PreferencesDialog::verifyDeriveTrackLanguageSettings() {
   }
 
   return true;
+}
+
+bool
+PreferencesDialog::verifyDeriveHearingImpairedFlagSettings() {
+  auto reText = ui->leMDeriveHearingImpairedFlagRE->text();
+  QRegularExpression re{reText, QRegularExpression::CaseInsensitiveOption};
+
+  if (!reText.isEmpty() && re.isValid())
+    return true;
+
+  Util::MessageBox::critical(this)
+    ->title(QY("Invalid settings"))
+    .text(QY("The value for deriving the 'hearing impaired' flag from file names must be a valid regular expression."))
+    .exec();
+
+  return false;
 }
 
 bool
@@ -1527,6 +1571,7 @@ PreferencesDialog::verifyRunProgramConfigurations() {
 void
 PreferencesDialog::accept() {
   if (   verifyDeriveTrackLanguageSettings()
+      && verifyDeriveHearingImpairedFlagSettings()
       && verifyDeriveForcedDisplayFlagSettings()
       && verifyRunProgramConfigurations()) {
     rememberCurrentlySelectedPage();