+++ src/mkvtoolnix-gui/merge/track.h
@@ -93,6 +93,7 @@ protected:
   void setDefaultsBasics();
   void setDefaultsMuxThis();
   void setDefaultsForcedDisplayFlag();
+  void setDefaultsHearingImpairedFlag();
   void setDefaultsDisplayDimensions();
   void setDefaultsLanguage(mtx::bcp47::language_c const &languageDerivedFromFileName);
   void setDefaultsColor();