+++ src/mkvtoolnix-gui/util/settings_names.h
@@ -36,7 +36,9 @@ char const * const s_valCommandLine                                             
 char const * const s_valDefaultAdditionalMergeOptions                           = "defaultAdditionalMergeOptions";
 char const * const s_valDefaultAudioTrackLanguage                               = "defaultAudioTrackLanguage";
 char const * const s_valDefaultChapterLanguage                                  = "defaultChapterLanguage";
+char const * const s_valDefaultDeriveHearingImpairedFlagFromFileNames           = "defaultDeriveHearingImpairedFlagFromFileNames";
 char const * const s_valDefaultDeriveSubtitlesForcedFlagFromFileNames           = "defaultDeriveSubtitlesForcedFlagFromFileNames";
+char const * const s_valDefaultRegexForDerivingHearingImpairedFlagFromFileNames = "defaultRegexForDerivingHearingImpairedFlagFromFileNames";
 char const * const s_valDefaultRegexForDerivingSubtitlesForcedFlagFromFileNames = "defaultRegexForDerivingSubtitlesForcedFlagFromFileNames";
 char const * const s_valDefaultSubtitleCharset                                  = "defaultSubtitleCharset";
 char const * const s_valDefaultSubtitleTrackLanguage                            = "defaultSubtitleTrackLanguage";